fix: address PR #118 code review findings (P0 + P1)
Fixes applied:
- Skeleton.tsx: migrated from RN Animated to react-native-reanimated, added dynamic shimmer width via onLayout
- Tooltip.tsx: migrated from RN Animated to react-native-reanimated (useSharedValue + withTiming + runOnJS)
- metal-holding-service.ts: added input validation (name, weight, price, purity) before DB write
- useMetalHoldings.ts: extracted business logic to metal-holding-calculations.ts, fixed loading stuck on error, added purchasePriceInPref to MetalHolding, extracted PERCENTAGE_MULTIPLIER constant
- metal-holding-calculations.ts: NEW file with pure computation functions extracted from hook
- tasks.md: fixed migration filename 024→034, corrected total task count 22→21
