+++ apps/mobile/hooks/useMetalHoldings.ts
@@ -18,7 +18,20 @@ import { useEffect, useMemo, useState } from "react";
 import { Q } from "@nozbe/watermelondb";
 
 import { Asset, AssetMetal, database, type CurrencyType } from "@astik/db";
-import { convertCurrency, getMetalPriceUsd } from "@astik/logic";
+import { convertCurrency } from "@astik/logic";
+
+import {
+  enrichHolding,
+  groupAndSortHoldings,
+  computePortfolioSplit,
+  joinAssetsWithMetals,
+  PERCENTAGE_MULTIPLIER,
+} from "../services/metal-holding-calculations";
+import type {
+  MetalHolding,
+  PortfolioSplit,
+  ProfitLoss,
+} from "../services/metal-holding-calculations";
 
 import { useMarketRates } from "./useMarketRates";
 import { usePreferredCurrency } from "./usePreferredCurrency";
@@ -27,41 +40,11 @@ import { usePreferredCurrency } from "./usePreferredCurrency";
 // Types
 // ---------------------------------------------------------------------------
 
-/** A metal holding enriched with computed values for display */
-interface MetalHolding {
-  /** The parent Asset record */
-  readonly asset: Asset;
-  /** The child AssetMetal record */
-  readonly assetMetal: AssetMetal;
-  /** Current market value in user's preferred currency */
-  readonly currentValue: number;
-  /** Current market value in USD */
-  readonly currentValueUsd: number;
-  /** Profit/loss as a percentage ((current - purchase) / purchase * 100) */
-  readonly profitLossPercent: number;
-  /** Absolute profit/loss amount in preferred currency */
-  readonly profitLossAmount: number;
-}
-
 /** Summary for a single metal type */
 interface MetalTypeSummary {
   readonly totalValue: number;
   readonly percentage: number;
   readonly itemCount: number;
-}
-
-/** Portfolio split between metal types */
-interface PortfolioSplit {
-  readonly gold: MetalTypeSummary;
-  readonly silver: MetalTypeSummary;
-}
-
-/** Aggregate profit/loss for the entire portfolio */
-interface ProfitLoss {
-  /** Absolute profit/loss amount in preferred currency */
-  readonly amount: number;
-  /** Profit/loss as a percentage */
-  readonly percent: number;
 }
 
 interface UseMetalHoldingsResult {
@@ -79,149 +62,6 @@ interface UseMetalHoldingsResult {
   readonly portfolioSplit: PortfolioSplit;
   /** Whether the data is still loading */
   readonly isLoading: boolean;
-}
-
-// ---------------------------------------------------------------------------
-// Internal Types (raw DB data before enrichment)
-// ---------------------------------------------------------------------------
-
-interface RawHolding {
-  readonly asset: Asset;
-  readonly assetMetal: AssetMetal;
-}
-
-// ---------------------------------------------------------------------------
-// Helper Functions
-// ---------------------------------------------------------------------------
-
-/**
- * Joins assets with their corresponding asset_metals records.
- * Each asset is expected to have exactly one asset_metal child.
- */
-function joinAssetsWithMetals(
-  assets: readonly Asset[],
-  assetMetals: readonly AssetMetal[]
-): readonly RawHolding[] {
-  const metalsByAssetId = new Map<string, AssetMetal>();
-
-  for (const metal of assetMetals) {
-    metalsByAssetId.set(metal.assetId, metal);
-  }
-
-  const holdings: RawHolding[] = [];
-
-  for (const asset of assets) {
-    const assetMetal = metalsByAssetId.get(asset.id);
-    if (assetMetal) {
-      holdings.push({ asset, assetMetal });
-    }
-  }
-
-  return holdings;
-}
-
-/**
- * Enriches a raw holding with computed market values and profit/loss.
- */
-function enrichHolding(
-  raw: RawHolding,
-  latestRates: NonNullable<Parameters<typeof getMetalPriceUsd>[1]>,
-  preferredCurrency: CurrencyType
-): MetalHolding {
-  const pricePerGramUsd = getMetalPriceUsd(
-    raw.assetMetal.metalType,
-    latestRates
-  );
-  const currentValueUsd = raw.assetMetal.calculateValue(pricePerGramUsd);
-  const currentValue =
-    preferredCurrency === "USD"
-      ? currentValueUsd
-      : convertCurrency(currentValueUsd, "USD", preferredCurrency, latestRates);
-
-  const purchasePriceInPref =
-    raw.asset.currency === preferredCurrency
-      ? raw.asset.purchasePrice
-      : convertCurrency(
-          raw.asset.purchasePrice,
-          raw.asset.currency,
-          preferredCurrency,
-          latestRates
-        );
-
-  const profitLossAmount = currentValue - purchasePriceInPref;
-  const profitLossPercent =
-    purchasePriceInPref > 0
-      ? (profitLossAmount / purchasePriceInPref) * 100
-      : 0;
-
-  return {
-    asset: raw.asset,
-    assetMetal: raw.assetMetal,
-    currentValue,
-    currentValueUsd,
-    profitLossPercent,
-    profitLossAmount,
-  };
-}
-
-/**
- * Sort comparator: newest purchase date first (descending).
- * FR-024: Holdings sorted by purchase date descending.
- */
-function sortByPurchaseDateDesc(a: MetalHolding, b: MetalHolding): number {
-  return b.asset.purchaseDate.getTime() - a.asset.purchaseDate.getTime();
-}
-
-/**
- * Group enriched holdings by metal type and sort by purchase date desc.
- */
-function groupAndSortHoldings(holdings: readonly MetalHolding[]): {
-  gold: MetalHolding[];
-  silver: MetalHolding[];
-} {
-  const gold: MetalHolding[] = [];
-  const silver: MetalHolding[] = [];
-
-  for (const holding of holdings) {
-    if (holding.assetMetal.metalType === "GOLD") {
-      gold.push(holding);
-    } else if (holding.assetMetal.metalType === "SILVER") {
-      silver.push(holding);
-    }
-  }
-
-  gold.sort(sortByPurchaseDateDesc);
-  silver.sort(sortByPurchaseDateDesc);
-
-  return { gold, silver };
-}
-
-/**
- * Compute portfolio split percentages.
- */
-function computePortfolioSplit(
-  goldHoldings: readonly MetalHolding[],
-  silverHoldings: readonly MetalHolding[],
-  totalValue: number
-): PortfolioSplit {
-  const goldTotal = goldHoldings.reduce((sum, h) => sum + h.currentValue, 0);
-  const silverTotal = silverHoldings.reduce(
-    (sum, h) => sum + h.currentValue,
-    0
-  );
-
-  return {
-    gold: {
-      totalValue: goldTotal,
-      percentage: totalValue > 0 ? (goldTotal / totalValue) * 100 : 0,
-      itemCount: goldHoldings.length,
-    },
-    silver: {
-      totalValue: silverTotal,
-      percentage: totalValue > 0 ? (silverTotal / totalValue) * 100 : 0,
-      itemCount: silverHoldings.length,
-    },
-  };
 }
 
 // ---------------------------------------------------------------------------
@@ -269,7 +109,9 @@ export function useMetalHoldings(): UseMetalHoldingsResult {
         setAssets(result);
       },
       error: (err: unknown) => {
-        console.error("Error observing metal assets:", err);
+        // TODO: Replace with structured logging when logging infrastructure is added
+        console.error("[useMetalHoldings] Error observing metal assets:", err);
+        setDataLoading(false);
       },
     });
 
@@ -287,7 +129,8 @@ export function useMetalHoldings(): UseMetalHoldingsResult {
         setDataLoading(false);
       },
       error: (err: unknown) => {
-        console.error("Error observing asset metals:", err);
+        // TODO: Replace with structured logging when logging infrastructure is added
+        console.error("[useMetalHoldings] Error observing asset metals:", err);
         setDataLoading(false);
       },
     });
@@ -322,25 +165,17 @@ export function useMetalHoldings(): UseMetalHoldingsResult {
     // 3. Group and sort (FR-024: newest first)
     const { gold, silver } = groupAndSortHoldings(enriched);
 
-    // 4. Compute aggregates
+    // 4. Compute aggregates (using pre-computed purchasePriceInPref from enrichHolding)
     const totalValue = enriched.reduce((sum, h) => sum + h.currentValue, 0);
-    const totalPurchasePrice = enriched.reduce((sum, h) => {
-      const purchaseInPref =
-        h.asset.currency === preferredCurrency
-          ? h.asset.purchasePrice
-          : convertCurrency(
-              h.asset.purchasePrice,
-              h.asset.currency,
-              preferredCurrency,
-              latestRates
-            );
-      return sum + purchaseInPref;
-    }, 0);
+    const totalPurchasePrice = enriched.reduce(
+      (sum, h) => sum + h.purchasePriceInPref,
+      0
+    );
 
     const profitLossAmount = totalValue - totalPurchasePrice;
     const profitLossPercent =
       totalPurchasePrice > 0
-        ? (profitLossAmount / totalPurchasePrice) * 100
+        ? (profitLossAmount / totalPurchasePrice) * PERCENTAGE_MULTIPLIER
         : 0;
 
     // 5. Portfolio split
